Fix spell and item image scaling and custom image handling
- Corrected scaling calculations for spell and item images to account for image scale factors in `SpellButton`, `ItemSlot`, and spell rendering logic.
- Added support for persisting and copying custom item images in `Item` and `ItemFactory`.
- Improved spell description merging in Lua to prioritize custom spell fields.
- Updated test spell data and spell list for consistency.
- Replaced manual setter in `ItemSlot` with Lombok's `@Setter` annotation for `inQuickSlot`.
- Minor code cleanup and bug fixes related to image and spell handling.

+++ RemixedDungeon/src/main/java/com/nyrds/pixeldungeon/mechanics/spells/Spell.java
@@ -169,7 +169,7 @@ public class Spell implements NamedEntityKind {
         var spellImage = new Image(texture);
 
         spellImage.frame(TextureCache.getFilm(texture, imageSize, imageSize).get(getImage()));
-        spellImage.setScale(16/imageSize);
+        spellImage.setScale((float) 16 /imageSize);
 
         return spellImage;
     }
